+++ tests/DefinitionSet/IntersectionTest.php
@@ -8,6 +8,7 @@ use Innmind\Math\{
     Algebra\Number,
     Exception\OutOfDefinitionSet,
 };
+use Innmind\Immutable\SideEffect;
 use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 
 class IntersectionTest extends TestCase
@@ -45,12 +46,15 @@ class IntersectionTest extends TestCase
             ),
         );
 
-        $this->assertNull($set->accept(Number::of(2)));
+        $this->assertInstanceOf(
+            SideEffect::class,
+            $set->accept(Number::of(2))->unwrap(),
+        );
 
         $this->expectException(OutOfDefinitionSet::class);
         $this->expectExceptionMessage('0.1 ∉ {1;2}∩{2;5}');
 
-        $set->accept(Number::of(0.1));
+        $_ = $set->accept(Number::of(0.1))->unwrap();
     }
 
     public function testUnion()