+++ tests/DefinitionSet/NaturalNumbersExceptZeroTest.php
@@ -8,6 +8,7 @@ use Innmind\Math\{
     Algebra\Number,
     Exception\OutOfDefinitionSet,
 };
+use Innmind\Immutable\SideEffect;
 use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 
 class NaturalNumbersExceptZeroTest extends TestCase
@@ -33,12 +34,15 @@ class NaturalNumbersExceptZeroTest extends TestCase
     {
         $set = Set::naturalNumbersExceptZero();
 
-        $this->assertNull($set->accept(Number::of(1)));
+        $this->assertInstanceOf(
+            SideEffect::class,
+            $set->accept(Number::of(1))->unwrap(),
+        );
 
         $this->expectException(OutOfDefinitionSet::class);
         $this->expectExceptionMessage('0.1 ∉ ℕ*');
 
-        $set->accept(Number::of(0.1));
+        $_ = $set->accept(Number::of(0.1))->unwrap();
     }
 
     public function testUnion()