+++ tests/DefinitionSet/RangeTest.php
@@ -8,6 +8,7 @@ use Innmind\Math\{
     Algebra\Number,
     Exception\OutOfDefinitionSet,
 };
+use Innmind\Immutable\SideEffect;
 use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 
 class RangeTest extends TestCase
@@ -47,12 +48,15 @@ class RangeTest extends TestCase
     {
         $set = Set::inclusiveRange(Number::of(1), Number::of(2));
 
-        $this->assertNull($set->accept(Number::of(1)));
+        $this->assertInstanceOf(
+            SideEffect::class,
+            $set->accept(Number::of(1))->unwrap(),
+        );
 
         $this->expectException(OutOfDefinitionSet::class);
         $this->expectExceptionMessage('0.1 ∉ [1;2]');
 
-        $set->accept(Number::of(0.1));
+        $_ = $set->accept(Number::of(0.1))->unwrap();
     }
 
     public function testUnion()